+++ covalent_ui/app/core/config.py
@@ -24,7 +24,8 @@ from typing import List, Union
 from dotenv import find_dotenv, load_dotenv
 from pydantic import AnyHttpUrl, BaseSettings, validator
 
-load_dotenv()
+HOME_PATH = os.environ.get("XDG_CACHE_HOME") or (os.environ["HOME"] + "/.cache")
+load_dotenv(f"{HOME_PATH}/covalent/.env")
 
 
 class Settings(BaseSettings):
@@ -43,7 +44,6 @@ class Settings(BaseSettings):
 
     class Config:
         case_sensitive = True
-        env_file = ".env"
 
 
 settings = Settings()